requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
coverage==7.6.4
distlib==0.3.9
Django==3.2.23
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
exceptiongroup==1.2.2
filelock==3.16.1
identify==2.6.1
idna==3.10
iniconfig==2.0.0
nodeenv==1.9.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
psycopg2-binary==2.9.10
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
